+++ CHANGELOG.md
@@ -41,3 +41,9 @@
             
     2.1.6
         - Add headers to respondWithObject request
+
+    3.0.0
+        - auditor is first class citizen instead of rawResponseCaller which no longer exists
+        - cache has cacheIgnoreFields option array so you can have it ignore fields on your request for cache keying
+        - fixed a cache key issue
+        - child has ability to have an extendRequest function which runs on every request before sent into parent